buildOpal.sh

+++ buildOpal-backup.sh
@@ -164,8 +164,11 @@ then
 	rm -r www
 	mkdir www
 
+
+
 	#Prod specific build phase, this includes gulp taks + using proper index.html file
 	if [ "$DEST" = "prod" ]; then
+
 		#Run Gulp tasks in order to compress and bundle all needed files
 		cd $WORKING_DIR
 
@@ -214,13 +217,46 @@ then
 		echo ""
 		echo ""
 		echo "Removing intermediate build folder..."
-	
-		#You can now remove the dest folder 
+
+		#You can now remove the dest folder
 		rm -r $WORKING_DIR/dest
 
 
+#		echo ""
+#		echo ""
+#		echo "Gulp build tasks were successful!"
+#		echo ""
+#		echo ""
+#		echo "Copying over other dependencies to dest folder..."
+#
+#		#Copy over language directory
+#		cp -a $WORKING_DIR/www/Languages/. $WORKING_DIR/dest/Languages
+#
+#		#Copy over fonts directory to root
+#		cp -a $WORKING_DIR/www/fonts/. $WORKING_DIR/dest/fonts
+#
+#		#Copy over fonts directory to vendor for OnsenUI
+#		cp -a $WORKING_DIR/www/fonts/. $WORKING_DIR/dest/vendor/fonts
+#
+#		cp $WORKING_DIR/www/lib/bower_components/bootstrap/dist/css/bootstrap.min.css.map $WORKING_DIR/dest/vendor
+#
+#		#Grab from dest folder and move them the production environment
+#		cp -a $WORKING_DIR/dest/. $TARGET_DIR/www
+#
+#		echo ""
+#		echo ""
+#		echo "Copying completed successfully..."
+#
+#		echo ""
+#		echo ""
+#		echo "Removing intermediate build folder..."
+#
+#		#You can now remove the dest folder
+#		rm -r $WORKING_DIR/dest
+
+
 	else
-		#Just transfer everything over if not in production mode 
+		#Just transfer everything over if not in production mode
 		cp -a $WORKING_DIR/www/. $TARGET_DIR/www
 		rm www/karma.conf.js
 		rm www/package.json
@@ -230,37 +266,51 @@ then
         fi
 	fi
 
-	#TODO: UPDATE VERSION # IN CONFIG.XML + Build.Gradle file 
+	#TODO: UPDATE VERSION # IN CONFIG.XML + Build.Gradle file
 	# IMPORTANT IMPORTANT: Use always Gradle version 2.2.2   DO NOT UPGRADE TO A HIGHER VERSION. IT WON'T WORK
 	# ALWAYS MAKE SURE THAT THIS LINE IS IN Build.Gradle file: classpath 'com.android.tools.build:gradle:2.2.2'
 
 	echo ""
 	echo ""
+	echo "*********************************************"
+	echo "Copying over config.xml to build destination..."
+	echo "*********************************************"
 	echo ""
 	echo "Copying over config.xml to build destination..."
-	# $version_android = 
+	# $version_android =
     # sed -e "s/version=\"[0-9.]*\" android-versionCode=\"[0-9]*\"/version=\"${version}\" android-versionCode=\"${version_android}\"" $TARGET_DIR/config.xml
 	#Remove current config.xml and replace with current one
-	rm $TARGET_DIR/config.xml
+	if [ -f "$TARGET_DIR/config.xml" ]; then
+	    rm $TARGET_DIR/config.xml
+	fi
 	cp $WORKING_DIR/config.xml $TARGET_DIR
 
+	if [ -f "$TARGET_DIR/package.json" ]; then
+        rm $TARGET_DIR/package.json
+	fi
 	#Move to target directory
 	cd $TARGET_DIR
 
 	echo ""
 	echo ""
+	echo "*********************************************"
+	echo "*********************************************"
 	echo ""
 	echo "All checks have been met... time to build...."
 	echo ""
+	echo "*********************************************"
+	echo "*********************************************"
 	echo ""
 	echo ""
 
 	#All checks have been met... time to build...
     cordova build -verbose
 
-	
+
 	echo ""
+	echo "*********************************************"
 	echo "Build was successful!"
+	echo "*********************************************"
 	echo ""
 
 	if [ "$DEST" = "prod" ]; then
